Reduce TOTAL_MEMORY to 5MiB
Since memory growth was enabled in #155, it seems prudent to minimize
initial memory footprint.

Prior to this, with `TOTAL_MEMORY` set to 1GiB, pyodide pre-allocates a
whole gigabyte of memory.

I could see the difference in memory usage quite easily in the developer
tools memory tab.

I've done some rudimentary testing with developer tools and the console,
and can confirm that doing things like allocating large byte arrays
dynamically grows the memory up to around the ~2GiB browser limit.

I discovered that if you set `Module.TOTAL_MEMORY` in `pyodide.js`, then
this overrides the compiled-in initial memory allocation. It'd be
awesome to have this be application-overrideable similar to how
`languagePluginUrl` is today. I'd be happy to put together a commit for
that if there's interest.