Recycling hiwire keys
hiwire is frequenly used to interchange values and objects between Python and JavaScript. Even when objects are deleted from hiwire.objects after usage, their key is never used again, and the offset is increased all the time until it might overflow when the site is executed for a very long time and with lots of interchanges.

This little improvement may avoid an overflow by re-using keys which already had been used before, using a list of recycled keys.