Remove external dependencies from file_packager.py
With the change, if `file_packager.py` is invoked without the `lz4`
configuration option, then it does not use any external dependencies,
making it easier for downstream users to package their python libraries.

This commit also adds documentation for how to package third-party
python libraries for use with Pyodide.

Fixes #808