Remove --host argument in pyodide_build
Instead, we assume that the currently running python is the host. This
simplifies the build script slightly and avoids making assumptions like
the binary is `sys.prefix`/bin/python3. Since it is easy for the user to
choose which python to use to run the scripts, there is no loss in
configurability.