Refactor buildall.py
The overarching idea of the refactoring is to build a global picture of
the dependency tree before compiling. The immediate benefit of this
refactoring is that PYODIDE_PACKAGES no longer needs manually including
the package dependencies.

Further, the refactor is done with two additional features in mind, to
be added in the future:

 - Instead of specifying skip_host in meta.yaml, a package can specify
   build dependencies, and skip_host is True iff it is not a build
   dependency of any other package

 - Build packages in parallel via a threadpool

In particular, the latter is difficult with the original recursive
approach.

The code is not the most efficient; it goes through the list of packages
multiple times. This script will be far from being the bottleneck, so I
prefer slightly more readable code than highly efficient ones.