Don't write our own instantiateWasm code
The default version works well and pretty much does the same thing, and
writing our own requires us to keep it up to date with what the default
version does. In particular, it breaks with emscripten 1.38.35.

The "customized" instantiateWasm code was added pretty early on in the
project, when emscripten was possibly less mature.