Removing lz4_c patch
We use lz4 to compress files in our virtual filesystem. The lz4_c patch
patches emscripten to use a wasm version of lz4 instead of the default
javascript version.

I don't think this was ever supposed to work, and it definitely does not
work with emscripten 1.38.35. The reason is that we use lz4 to manage
the filesystem, which is loaded before the emscripten runtime is
initialized, and we are not supposed to call wasm functions before the
runtime is initialized.

The patch was originally included for performance reasons, but it
appears that there is in fact not much measurable performance
difference - the decompression algorithm is fairly straightforward and
presumably not difficult to JIT well.