Disable failing test
This starts failing in 1.38.35. In fact, the test would never have
passed, but before 1.38.34, this failure is pre-empted by an earlier
failure that causes the test to skip.

Indeed, in `test_largefile.py`, the `setupModule` part checks if the
operating system is able to deal with large files, and skips the test if
it is not. With 1.38.34 it cannot, while in 1.38.35, it can. However,
while the "operating system" can now deal with large files, the underlying
emscripten filesystem MEMFS cannot, and it never could. We can reproduce
the failure with 1.38.34 if we replace 2_500_000_000 in
test_largefile.py with 2 ** 31 - 1, which is a filesize small enough for
the OS but too large for the FS.