Fix cpython configure
This issue starts with emscripten 1.38.38.

This effectively reverts the upstream commit
https://github.com/emscripten-core/emscripten/commit/08d5e0c36d662104ac27432d5fc872423b9e578f

./configure scripts work by trying to compile snippets of code to see if
certain functions are present. With fastcomp, the compilation usually
goes

C -> asm.js -> wasm

Previously, during configuration, we just compile to asm.js and call it
a day. In 1.38.38, it compiles all the way down to wasm. It turns out
that certain functions (e.g. copysign, acos) do exist, but compiling
down to wasm makes configure think they don't exist. This patch is a
temporary hack to retain the old behaviour. I don't think it is worth
the effort to wrestle with this (most likely fruitlessly) if we intend
to replace fastcomp with upstream llvm, which directly compiles to wasm.
The compilation pathway will be completely different and configure will
most likely behave in independent ways.