Streamline hiwire_get_dtype (#956)
Implementation of hiwire_get_dtype is split across two files. The reason seems to be that get_dtype returns two values: format and size. I changed it to take pointers to write the results into. This makes the code shorter and easier to understand (someone reading the files would have to search the source to find out that those constants are only used in two spots).