Fix pywasmcross race condition (#1011)
When building packages in parallel, it is possible that the symlink is
created by a different thread between the check and the linking, which
causes the build to fail. This has happened multiple times in CI.

Instead, just try to create the symlink and pass if it already exists,
which is the pythonic way of doing this.