Refactor conftest.py
This eliminates a lot of duplicate code using selenium's
driver.execute_async_script.

Error handling is manual, by returning a discriminator. The "proper"
solution should be to use selenium's JavascriptException type, but it
doesn't seem to support Chrome, and we get a generic WebDriverException
that is less useful.

This ends up simpliflying the error output, but there is still work to
be done on that front.