Refactor loadPackage
From the API point of view, this removes the messageCallback and
errorCallback arguments in loadPackage, runPythonAsync etc. These are
used for logging. Previously, any message will be written to the
console, and then {message,error}Callback is called with the message as
an argument. In most of our use cases, these callbacks do nothing.

If we were to reintroduce this feature, I think we should introduce a
logger interface, and users can supply custom loggers. This lets them
choose to not output to stdout/stderr as well.

Under the hood, the original motivation of this patch was to get rid of
using packageCounter to determine whether all packages have been loaded.
This is prone to error if we change the way we load packages (which I
attempted in a different branch until I encountered an emscripten bug).

One way or another, I ended up rewriting a lot of the code into
something with a clearer control flow imo.