Rework package loading algorithm
The main difference in end result in we no longer produce a hard error
when errors occur during the dependency resolution stage; we simply
ignore the offending packages. We still error if something goes wrong
during package loading, as the system may be left in a broken state.

The error messages produced are, however, slightly different. Now if a
package is loaded from a custom url then the default channels, we no
longer print it as an error; we assume the user is trying to override a
dependency. Since we don't act on these errors anyway, this doesn't
affect the API.

Smaller changes include changing to a recursive DFS algorithm and
turning toLoad into a Map