made os.waitpid be a no-op
os.waitpid is called by various test functions. Unfortunately it does a syscall, which in recent versions of emscripten calls abort on the wasm runtime.
It should never be used outside the tests, because we don't support multi-process yet. As such it should be okay to make it a no op (and print a warning)