Made scipy/CLAPACK not have any missing functions or type conflicts
this mostly involved sorting out the scipy fortran wrappers to actually call
direct to lapack functions rather than using wrapped functions which don't exist
in clapack.

I also brought back the LAPACK functions in scipy/linalg/deprecated in the CLAPACK versions,
and removed the deprecated versions from the scipy build. Scipy just assumes you can define things twice and
be cavalier with signatures which means it will work with modern lapack and provide those functions. However
on webassembly, that doesn't work, because in some situations, if something is defined twice, webassembly just
dumps one of the functions silently and aborts on all calls to it. You can spot this by looking for
mismatch:fnname in the wasm modules.