Use async functions for languageLoaderPlugin
This allows us to use more conventional "modern" javascript.

In the future, I hope to make languageLoaderPlugin a function instead of
a promise (which would be very easy given this commit), so that users
can supply a custom Module object. One useful application would be to
supply stdout/stderr handlers. This would be a breaking change so I'll
leave it for antoher time.