Refactor core test suite
The goal of this PR is to run more tests in the core test suite.
In order to do this, I converted python_tests.txt into a yaml file
so that we can have slightly more control over what goes on. The
original xfail behavior we now get with the xfail key, but we can
also skip specific test patterns with a skip key.

I then went through and fairly carefully skipped as few tests was as
reasonable. I made .so files for _test_sqlite3 and _test_ctypes so that
they can be tested as well.