Remove bad ctypes patch
This resolves #2330.

As tiran helpfully points out, the Don't-dereference-function-pointer
patch is wrong. Most likely it was always wrong, and I added it thinking
that I was fixing a problem that never existed or that was fixed by some
other change (possibly to the libffi port) that I didn't test separately.
Removing this patch makes many more of the ctypes tests pass.