Remove pandas docstrings
Local patch of https://github.com/pandas-dev/pandas/pull/46934

I don't expect this to make a user-noticeable improvement, but it seems
worth doing regardless. Locally I see a reduction from 4546988 for the
whole wheel to 4545234 (so just under 1k) although some of the .so's
inside are a few kB smaller.