docs for new packages: add info how to try out locally built wheel
After managing to build the first wheel for a new package (obspy)
locally, I was kinda sitting there with the compiled wheel and trying to
figure out how to try it out locally.
I took a bit to figure out these steps and maybe they might help other
people too.

If there's a better way to do this, like auto-add to the package.json or
anything (without building the full pyodide build), this commit should
be improved obviously. Total newb to pyodide.