ENH Add js_id attribute to JsProxy for use as a dictionary key (#2515)
It's useful to be able to hash on object identity. For instance in Python
if we want to store objects in a dictionary using object identity to
store/access obj we use `id(obj)` as the index. This adds `js_id` as an
attribute to `JsProxy` so that `jsobj.js_id` can be used as a dictionary key
to key on identity of the JavaScript object.