Don't use ** for fetch_extra_kwargs
It's an antipattern to use ** into your own private APIs.
If a function is expecting a dictionary, we should change their signature
so they expect a dictionary. Otherwise, some day we might add a key to
the dictionary which happens to match the name of some other argument
in one of our private APIs and cause logic errors.