MAINT Consolidate `ctx` and `fetch_extra_kwargs` into `Transaction` (#2564)
This is more cleanup in micropip. I moved `ctx` and `fetch_extra_kwargs` into `Transaction`
properties so that we don't have to pass them through everywhere as separate arguments.
I also switch to creating the `Transaction` in `install` so that we don't have to pass the arguments 
down into `gather_requirements` separately.