Add pyodide_build create_xbuildenv and install_xbuildenv (#2550)
This adds pyodide_build command create_xbuildenv which creates a
crossbuild environment (from a copy of Pyodide where scipy has been
build) and install_xbuildenv which installs the cross build environment
into a fresh copy of Pyodide.

I successfully installed the xbuild environment into a fresh checkout of
Pyodide then built statsmodels and scikit-learn in isolation, without 
building the Python interpreter, numpy, or scipy. I dumped the generated
wheels into a copy of Pyodide downloaded from CI, and was able to import
and use them as normal.

The size of the xbuild environment is 1.5 megabytes, of which 1.2 megabytes
is Python headers.

In a subsequent PR, we can update the CI to automatically upload these
to aws s3 and then install the environment from there.