ENH Use `dist-info` + `importlib.metadata` to store and retrieve package metadata (#2572)
This switches to using the file system to store the information about packages
and using importlib.metadata to retrieve it.
This has two related benefits:

    We don't have to separately maintain our own separate state about what
    we've installed.

    We are guaranteed to agree with the Python stdlib about whether or not a
    package is installed, and if so which version is installed. Since the state is
    maintained in only one place, there is no chance of it diverging.

According to the packaging specs, if the package is from a url we should put
that into a file here:
packaging.python.org/en/latest/specifications/direct-url/#direct-url
Other than that, we should set the INSTALLER to pyodide.loadPackage
if the package was loaded with pyodide.loadPackage either directly or
indirectly via micropip. Otherwise set INSTALLER to micropip. That
way we can maintain the current behavior of micropip.list:

    if direct_url.json is present, then source is the url in direct_url.json
    otherwise, set source to pyodide if INSTALLER is pyodide.loadPackage
    and set it to micropip if INSTALLER is micropip.

Oddly enough, the packaging specs suggest no place to put the source
index, so it seems that according to the specs if a package was installed
by name from some custom index, this info should not be recorded.