TST Use short pytest tracebacks (#2700)
By default, pytest uses the --tb=auto option with 'long' tracebacks for 
the first and last entry, but 'short' style for the other entries.

Since the last entry is typically a very long selenium function, it would
print all the source code of that function making it more difficult to find
relevant information. The first entry is the test function, and it's also not
very useful to print it in full.