Wrap exception wrapper logging in #ifdef DEBUG_F.
Currently, all Python exceptions that are converted to a JavaScript
exception are logged to the console as a waring. This can create quite
a bit of noise in the console when exceptions are expected.

This wraps the logging in `#ifdef DEBUG_F` so that we only log this in
debug builds.