Refactor hiwire.[ch] to divide functions declarations into better groups (#2847)
The functions inside `hiwire.h` and `hiwire.c` were defined in some inconsistent order. This PR shuffle things around so that they are clearly divided into multiple groups, one for each "namespace" (in the C sense):
- `hiwire_*`
- `JsObject_*`
- `JsArray_*`
- `JsString_*`
- `JsMap_*`
- `JaString_*`

This makes it easier to navigate inside the file, and will make it easier to add future functions "at the right place".