Fix doc for building package with pyodide_build (#2854)
In the old instructions to build package it was written that you should
run pyodide_build buildall from the packages/<package-name> folder.
However, the command that comes next assumes you are in the root
(like the previous commands, i.e. to create the meta.yml template).
Fixed by removing the reference to the folder.