Be more careful about different array variants (#2938)
This fixes a few problems introduced in #2907 (and some older issues). The slicing doesn't work correctly for `TypedArray` or `NodeList` and there are several possible routes to take with `TypedArray` so I think it's better to leave it not implemented for now. Also, `extend` and `+=` don't work for `TypedArray` and `NodeList`. This also gives a better error if you attempt `del a[0]` when `a` is a `TypedArray` or `NodeList`. The former have a fixed length, the latter are immutable.