Add browser polyfill for node url package.
This fixes the error:

```
Module not found: Error: Can't resolve 'url' in '/home/david/work/pybricks/pybricks-code/node_modules/pyodide'
BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
        - add a fallback 'resolve.fallback: { "url": require.resolve("url/") }'
        - install 'url'
If you don't want to include a polyfill, you can use an empty module like this:
        resolve.fallback: { "url": false }
```

when using webpack. This package was missed in #2468.